requirements.txt
opencv-contrib-python
numpy
requests
fastapi
uvicorn
pyodbc
paho-mqtt
ujson
python-multipart
